Split all routes to a separate file
